+++ app/views/hosts/edit.html.erb
@@ -4,4 +4,4 @@
     hash_for_toggle_manage_host_path(:id => @host.name), {:method => :put}) %>
 <% end -%>
 
-<%= render :partial => 'form' %>
+<%= render :partial => 'hosts/form' %>
