+++ lib/foreman/sti.rb
@@ -17,7 +17,14 @@ module Foreman
 
           alias_method_chain :new, :cast
         end
+        base.alias_method_chain :save, :type
       end
+    end
+    def save_with_type(*args)
+      self.class.instance_variable_set("@finder_needs_type_condition", :false) if self.type_changed?
+      value = save_without_type(*args)
+      self.class.instance_variable_set("@finder_needs_type_condition", :true)  if self.type_changed?
+      return value
     end
   end
 end
